Some minor changes
Uploaded the new version of Boa.py, but have not yet changed everything
in seraph to run on it. Next step for seraph.py is to fix all changes to
run with new Boa.py and than I can start coding the HTTP server to run
external api calls for use with javascript.